Rename to ``imod.util.regrid``
